requirements.txt
appdirs==1.4.4
attrs==19.3.0
black==19.10b0
cachetools==4.1.1
certifi==2020.6.20
cfgv==3.1.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
configparser==5.0.0
crayons==0.3.1
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.18.0
google-auth-httplib2==0.0.4
google-cloud==0.34.0
google-cloud-core==1.3.0
google-cloud-firestore==1.7.0
googleapis-common-protos==1.52.0
grpcio==1.30.0
httplib2==0.18.1
identify==1.4.20
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mypy==0.782
mypy-extensions==0.4.3
nodeenv==1.4.0
pathspec==0.8.0
pip-autoremove==0.9.1
pre-commit==2.5.1
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.9.8
pytz==2020.1
PyYAML==5.3.1
regex==2020.6.8
requests==2.24.0
rsa==4.6
selenium==3.141.0
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
uritemplate==3.0.1
urllib3==1.25.9
virtualenv==20.0.25
webdriver-manager==3.2.1
Werkzeug==1.0.1
